Refactor requests app
