+++ src/apps/requests/api/views/__init__.py
@@ -1,0 +1,1 @@
+from .request_view_set import RequestViewSet
