+++ src/apps/requests/models/__init__.py
@@ -1,0 +1,1 @@
+from .request import Request
